Fix linker issue related to fluid_file_test
