+++ src/utils/fluid_sys_glib.h
@@ -52,6 +52,8 @@ typedef gintptr  intptr_t;
 
 #include <glib/gstdio.h>
 
+#include "fluid_file.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -281,10 +283,6 @@ typedef GModule fluid_module_t;
 #endif /* LADSPA */
 
 /* File access */
-#define FLUID_FILE_TEST_EXISTS G_FILE_TEST_EXISTS
-#define FLUID_FILE_TEST_IS_REGULAR G_FILE_TEST_IS_REGULAR
-
-#define fluid_file_test(path, flags) g_file_test(path, flags)
 #define fluid_stat(_filename, _statbuf)   g_stat((_filename), (_statbuf))
 #if !GLIB_CHECK_VERSION(2, 26, 0)
     /* GStatBuf has not been introduced yet, manually typedef to what they had at that time:
